fix: lint

+++ course_learning/src/routes/dev-only/delete-all.ts
@@ -1,9 +1,5 @@
 import express, { Request, Response } from 'express';
-import { User } from '../../models/user';
 import {
-  NotAuthorizedError,
-  NotFoundError,
-  requireAuth,
   ResourcePrefix,
   EnvStage,
   BadRequestError,
@@ -23,6 +19,10 @@ router.delete(
 
     if (process.env.NODE_ENV === EnvStage.Prod) {
       throw new BadRequestError('This route is only available in development');
+    }
+
+    if (deleteKey !== 'CourseLearningDelete') {
+      throw new BadRequestError('Invalid delete key');
     }
 
     await sequelize.sync({ force: true });
